change type of account balance

+++ src/main/java/org/emuba/bankingemulation/dto/AccountDTO.java
@@ -3,19 +3,21 @@ package org.emuba.bankingemulation.dto;
 import lombok.Data;
 import org.emuba.bankingemulation.enums.TypeCurrency;
 
+import java.math.BigDecimal;
+
 @Data
 public class AccountDTO {
     private String accountNumber;
     private TypeCurrency currency;
-    private double balance;
+    private BigDecimal balance;
 
-    private AccountDTO(String accountNumber, TypeCurrency currency, double balance) {
+    private AccountDTO(String accountNumber, TypeCurrency currency, BigDecimal balance) {
         this.accountNumber = accountNumber;
         this.currency = currency;
         this.balance = balance;
     }
 
-    public static AccountDTO of(String accountNumber, TypeCurrency currency, double balance) {
+    public static AccountDTO of(String accountNumber, TypeCurrency currency, BigDecimal balance) {
         return new AccountDTO(accountNumber, currency, balance);
     }
