+++ src/main/java/org/emuba/bankingemulation/services/impl/AccountServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -64,7 +65,7 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     @Transactional
-    public synchronized void updateBalance(Long clientId, TypeCurrency currency, double newBalance) {
+    public synchronized void updateBalance(Long clientId, TypeCurrency currency, BigDecimal newBalance) {
         accountRepository.updateBalance(clientId, currency, newBalance);
     }
 }
